Feat: added dark mode (#147)
* Feat: Added support for dark theme

* Feat: added dark mode

* removed logs

* changed the button primary color

* changed the message theme color

* changed the tag color

+++ src/pages/api/v1/resource/delete/[resourceId].ts
@@ -15,7 +15,6 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     });
 
     if (findResource) {
-
       const videoLesson = await prisma.video.findUnique({
         where: {
           resourceId: Number(resourceId),
@@ -24,8 +23,6 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
           providerVideoId: true,
         },
       });
-
-
       if (videoLesson) {
         const serviceProviderResponse = await prisma?.serviceProvider.findFirst({
           where: {
@@ -46,7 +43,17 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
             serviceProvider
           );
         }
-        const [updateSeq, deleteCourse] = await prisma.$transaction([
+        const [updateSeq, deleteResource] = await prisma.$transaction([
+          prisma.$executeRaw`UPDATE Resource SET sequenceId = sequenceId - 1  WHERE sequenceId > ${findResource.sequenceId}  AND  chapterId = ${findResource.chapterId};`,
+          prisma.resource.delete({
+            where: {
+              resourceId: Number(resourceId),
+            },
+          }),
+        ]);
+      }
+      if (!videoLesson) {
+        const [updateSeq, deleteResource] = await prisma.$transaction([
           prisma.$executeRaw`UPDATE Resource SET sequenceId = sequenceId - 1  WHERE sequenceId > ${findResource.sequenceId}  AND  chapterId = ${findResource.chapterId};`,
           prisma.resource.delete({
             where: {
